+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Bug Fixes
 
+* (modules/coinswap)[\#153](https://github.com/irisnet/irismod/pull/151) Fix query not-existent reserve pool.
 * (modules/service) [\#152](https://github.com/irisnet/irismod/pull/152) Fix update service binding.
 
 ### Improvements